style(src) add src and include,move the files accordingly
